fix(接口自动化): 接口定义，sql请求没法执行；恢复编辑模块更新场景路径

+++ backend/src/main/java/io/metersphere/api/service/ApiModuleService.java
@@ -240,12 +240,14 @@ public class ApiModuleService extends NodeTreeService<ApiModuleDTO> {
                 if (apiDefinition != null && StringUtils.isNotBlank(apiDefinition.getModulePath())) {
                     StringBuilder path = new StringBuilder(apiDefinition.getModulePath());
                     List<String> pathLists = Arrays.asList(path.toString().split("/"));
-                    pathLists.set(request.getLevel(), request.getName());
-                    path.delete(0, path.length());
-                    for (int i = 1; i < pathLists.size(); i++) {
-                        path = path.append("/").append(pathLists.get(i));
+                    if (pathLists.size() > request.getLevel()) {
+                        pathLists.set(request.getLevel(), request.getName());
+                        path.delete(0, path.length());
+                        for (int i = 1; i < pathLists.size(); i++) {
+                            path = path.append("/").append(pathLists.get(i));
+                        }
+                        apiDefinition.setModulePath(path.toString());
                     }
-                    apiDefinition.setModulePath(path.toString());
                 }
             });
             batchUpdateApiDefinition(apiDefinitionResults);
